fix(venue): fiat config casts, forced_wire_status escape hatch, HTTP envelope
- Cast advance_seconds/recv_window manually instead of config()->integer(),
  which throws on the numeric-string a real .env produces
- Add forced_wire_status: an escape hatch for a wire vocabulary the enum
  doesn't model, echoed verbatim and never crediting — exercises the
  consumer's fail-closed arm
- Stop persisting forced_status into status: it's a read-time mask now,
  so clearing the override resumes the lazy advance instead of freezing it
- Add Completed/RefundFailed/PartialCreditStopped cases so every wire
  synonym the ticket lists is producible in both dialects
- Convert missing/malformed params on both fiat endpoints into the -1102
  fiat envelope instead of Laravel's default 422 body
- Align get-order-detail's data shape with the observed fixture:
  fiatCurrency, two-decimal amount, method/totalFee/createTime/updateTime,
  and drop pixcode once an order dies in a terminal failure state
- Strip banned issue/story references from src comments; record the
  dialect and HTTP-200-refusal decisions in ADR-0001

+++ app-modules/venue/config/venue-fiat.php
@@ -16,7 +16,7 @@ return [
     |
     */
 
-    'advance_seconds' => env('FAKE_BINANCE_FIAT_ADVANCE_SECONDS', 60),
+    'advance_seconds' => (int) env('FAKE_BINANCE_FIAT_ADVANCE_SECONDS', 60),
 
     /*
     |--------------------------------------------------------------------------
